+++ compiler-plugin-tests/src/test/java/io/ballerina/stdlib/graphql/compiler/ServiceValidationTest.java
@@ -1258,6 +1258,21 @@ public class ServiceValidationTest {
         assertErrorMessage(diagnostic, message, 34, 32);
     }
 
+    @Test(groups = "invalid")
+    public void testInvalidUsageDeprecatedDirective() {
+        String packagePath = "73_invalid_usages_of_deprecated_directive";
+        DiagnosticResult diagnosticResult = getDiagnosticResult(packagePath);
+        Assert.assertEquals(diagnosticResult.warningCount(), 2);
+        Iterator<Diagnostic> diagnosticIterator = diagnosticResult.warnings().iterator();
+
+        Diagnostic diagnostic = diagnosticIterator.next();
+        String message = getErrorMessage(CompilationDiagnostic.UNSUPPORTED_INPUT_FIELD_DEPRECATION, "Profile");
+        assertWarningMessage(diagnostic, message, 21, 12);
+
+        diagnostic = diagnosticIterator.next();
+        assertWarningMessage(diagnostic, message, 23, 9);
+    }
+
     private DiagnosticResult getDiagnosticResult(String packagePath) {
         Path projectDirPath = RESOURCE_DIRECTORY.resolve(packagePath);
         BuildProject project = BuildProject.load(getEnvironmentBuilder(), projectDirPath);
@@ -1279,6 +1294,12 @@ public class ServiceValidationTest {
 
     private void assertErrorMessage(Diagnostic diagnostic, String message, int line, int column) {
         Assert.assertEquals(diagnostic.diagnosticInfo().severity(), DiagnosticSeverity.ERROR);
+        Assert.assertEquals(diagnostic.message(), message);
+        assertErrorLocation(diagnostic.location(), line, column);
+    }
+
+    private void assertWarningMessage(Diagnostic diagnostic, String message, int line, int column) {
+        Assert.assertEquals(diagnostic.diagnosticInfo().severity(), DiagnosticSeverity.WARNING);
         Assert.assertEquals(diagnostic.message(), message);
         assertErrorLocation(diagnostic.location(), line, column);
     }