+++ compiler-plugin/src/main/java/io/ballerina/stdlib/graphql/compiler/diagnostics/DiagnosticCode.java
@@ -70,5 +70,6 @@ public enum DiagnosticCode {
     GRAPHQL_146,
     GRAPHQL_147,
     GRAPHQL_148,
-    GRAPHQL_149
+    GRAPHQL_149,
+    GRAPHQL_201
 }