+++ compiler-plugin/src/main/java/io/ballerina/stdlib/graphql/compiler/service/validator/ServiceValidator.java
@@ -55,6 +55,7 @@ import io.ballerina.stdlib.graphql.commons.types.TypeName;
 import io.ballerina.stdlib.graphql.compiler.Utils;
 import io.ballerina.stdlib.graphql.compiler.diagnostics.CompilationDiagnostic;
 import io.ballerina.stdlib.graphql.compiler.service.InterfaceEntityFinder;
+import io.ballerina.tools.diagnostics.DiagnosticSeverity;
 import io.ballerina.tools.diagnostics.Location;
 
 import java.util.ArrayList;
@@ -893,6 +894,11 @@ public class ServiceValidator {
                               recordTypeName);
             }
             for (RecordFieldSymbol recordFieldSymbol : recordTypeSymbol.fieldDescriptors().values()) {
+                boolean isDeprecated = recordFieldSymbol.deprecated();
+                if (isDeprecated) {
+                    addDiagnostic(CompilationDiagnostic.UNSUPPORTED_INPUT_FIELD_DEPRECATION,
+                            getLocation(recordFieldSymbol, location), recordTypeName);
+                }
                 validateInputType(recordFieldSymbol.typeDescriptor(), location, isResourceMethod);
             }
         }
@@ -972,12 +978,14 @@ public class ServiceValidator {
     }
 
     private void addDiagnostic(CompilationDiagnostic compilationDiagnostic, Location location) {
-        this.errorOccurred = true;
+        this.errorOccurred =
+                compilationDiagnostic.getDiagnosticSeverity() == DiagnosticSeverity.ERROR || this.errorOccurred;
         updateContext(this.context, compilationDiagnostic, location);
     }
 
     private void addDiagnostic(CompilationDiagnostic compilationDiagnostic, Location location, Object... args) {
-        this.errorOccurred = true;
+        this.errorOccurred =
+                compilationDiagnostic.getDiagnosticSeverity() == DiagnosticSeverity.ERROR || this.errorOccurred;
         updateContext(this.context, compilationDiagnostic, location, args);
     }